requirements.txt
pyhocon==0.3.58
pytest==6.2.5
gazpacho==1.1
selenium==4.0.0b4
requests==2.26.0
black==21.9b0
flake8==4.0.1
pandas==1.3.4
numpy>=1.17.3
python_dateutil>=2.7.3
pytz>=2017.3
undetected-chromedriver==3.0.3
random_user_agent==1.0.1
tqdm==4.62.3
scikit-learn==1.0.1
tensorflow==2.6.0
keras==2.6.0
keras-preprocessing==1.1.2
Pillow==8.4.0
matplotlib==3.4.3
mlflow==1.21.0
seaborn==0.11.2
notebook==6.4.6
jupyter_contrib_nbextensions==0.5.1
